Move tests for `setSelection` into separate file
